Move SaveSchemaWithDebounce into a hook
